Modified variable caloriesBurned type in the method in exercise, and implemented exerciseTest class

+++ src/main/java/it/unibo/workitout/model/workout/impl/Exercise.java
@@ -48,7 +48,7 @@ public class Exercise {
      * 
      * @return the calories burned as double value.
      */
-    public double calorieBurned(final int minutes) {
+    public double calorieBurned(final double minutes) {
         return this.calorieBurned * minutes;
     }
